requirements.txt
asgiref==3.5.2
black==22.3.0
boto3==1.24.17
botocore==1.27.17
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
fastapi==0.78.0
FastAPI-SQLAlchemy==0.2.1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.2
greenlet==1.1.2
h11==0.13.0
httplib2==0.20.4
idna==3.3
inflect==5.6.0
jmespath==1.0.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.37
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
